Overhaul of backend design to make up for many years of development without refactoring it for cleanliness [unstable build]
Big thing to note is distribution of logic from ElementPanel to a few new sub-classes, most notably ElementLoader which has all of the 'addElement' functions now. ElementPanel can be prompted for this object, though HandlePanels work the same as they route their commands through that interface automatically. Largely just simplifying some logic in the design and adding bespoke mutex locks for the drawList and clickList instead of both using the mutex of Panel; unfortunately, there is currently an error that appears in more complicated uses of this program than the in-built testing which I need to track down, so I wouldn't call this stable at the moment.
Will try to document this properly at some point, rough to go over it after all this time though I am impressed by some of the work I'd done.